pyproject.toml
[project]
name = "kog"
version = "0.1.0"
description = "Context-Aware CLI AI Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.1",
    "langchain-community>=0.3.14",
    "langchain-chroma>=0.2.0",
    "langchain-ollama>=0.2.2",
    "chromadb>=0.6.0",
    "pydantic>=2.10.0",
    "unstructured>=0.21.5",
    "markdown>=3.10.2",
    "langchain-text-splitters>=0.3.0",
    "langchain>=1.2.13",
    "langgraph>=1.0.1",
    "python-dotenv>=1.2.2",
    "pypdf>=6.9.1",
]

[project.scripts]
kog = "kog.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
